Move joi back to application where it should be
