+++ packages/common/src/validation/index.ts
@@ -86,10 +86,16 @@ export const transactionType = Joi.string().valid(
   ...Object.values(ETransactionType)
 );
 
+export const databaseName = Joi.string()
+  .trim()
+  .min(4)
+  .max(128)
+  .required()
+  .pattern(/^[a-z]+[_a-z0-9]+/i);
+
+export const merkleHeight = Joi.number().integer().positive().required();
+
 export const PERMISSION_DEFAULT_VALUE = Permission.policyStrict();
 
 export * from './collection';
 export * from './common';
-export * from './database';
-export * from './group';
-export * from './user';
