+++ packages/serverless/src/apollo/app/group.ts
@@ -1,11 +1,7 @@
 import {
-  JOI_GROUP_ADD_USER,
-  JOI_GROUP_CREATE,
-  JOI_GROUP_DETAIL,
-  JOI_GROUP_LIST_ALL,
-  JOI_GROUP_LIST_USER,
-  JOI_GROUP_REMOVE_USER,
-  JOI_GROUP_UPDATE,
+  databaseName,
+  groupDescription,
+  groupName,
   TGroupAddUserListRequest,
   TGroupAddUserListResponse,
   TGroupCreateRequest,
@@ -20,10 +16,12 @@ import {
   TGroupRemoveUserListResponse,
   TGroupUpdateRequest,
   TGroupUpdateResponse,
+  userName,
 } from '@zkdb/common';
 import { withTransaction } from '@zkdb/storage';
 import GraphQLJSON from 'graphql-type-json';
 
+import Joi from 'joi';
 import { Group } from '../../domain/use-case/group.js';
 import { gql } from '../../helper/common.js';
 import ModelGroup from '../../model/database/group.js';
@@ -97,6 +95,46 @@ export const typeDefsGroup = gql`
     ): Boolean
   }
 `;
+// Joi validation
+
+export const JOI_GROUP_CREATE = Joi.object<TGroupCreateRequest>({
+  databaseName,
+  groupName,
+  groupDescription: groupDescription(false),
+});
+
+export const JOI_GROUP_UPDATE = Joi.object<TGroupUpdateRequest>({
+  databaseName,
+  groupName,
+  newGroupName: groupName(false),
+  newGroupDescription: groupDescription(false),
+});
+
+export const JOI_GROUP_DETAIL = Joi.object<TGroupInfoDetailRequest>({
+  databaseName,
+  groupName,
+});
+
+export const JOI_GROUP_LIST_USER = Joi.object<TGroupListByUserRequest>({
+  databaseName,
+  userName,
+});
+
+export const JOI_GROUP_LIST_ALL = Joi.object<TGroupListAllRequest>({
+  databaseName,
+});
+
+export const JOI_GROUP_ADD_USER = Joi.object<TGroupAddUserListRequest>({
+  databaseName,
+  groupName,
+  listUser: Joi.array().items(Joi.string().required()).required(),
+});
+
+export const JOI_GROUP_REMOVE_USER = Joi.object<TGroupRemoveUserListRequest>({
+  databaseName,
+  groupName,
+  listUser: Joi.array().items(Joi.string().required()).required(),
+});
 
 // Query
 const groupListAll = publicWrapper<TGroupListAllRequest, TGroupListAllResponse>(
